requirements.txt
# ------------------------------
# 🧰 Core Utilities
# ------------------------------
pandas>=1.5.0

# ------------------------------
# 🔐 Data Generation & PII Utilities
# ------------------------------
faker==19.6.2

# hashlib is in Python standard library

# ------------------------------
# 🧠 NLP & Sentiment Analysis
# ------------------------------
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# ------------------------------
# 🔥 Spark Processing
# ------------------------------
pyspark==3.4.1

# ------------------------------
# ⚡ Real-Time Streaming
# ------------------------------
kafka-python==2.0.2

# ------------------------------
# 📊 Visualization
# ------------------------------
streamlit==1.29.0

# ------------------------------
# 📅 Orchestration (optional, Airflow)
# ------------------------------
apache-airflow==2.8.1

# ------------------------------
# 🧪 Testing Utilities (optional)
# ------------------------------
pytest>=7.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0